requirements.txt
smolagents
torch
transformers
Pillow
diffusers
timm
accelerate
scipy